+++ src/utils.js
@@ -1,0 +1,9 @@
+'use strict';
+
+var photosContainer = document.querySelector('.pictures');
+
+module.exports = {
+  isBottomReached: function() {
+    return photosContainer.getBoundingClientRect().bottom - window.innerHeight <= 60;
+  }
+};
